Split `qdrant::common::points` module (#5862)
