+++ src/actix/api/discovery_api.rs
@@ -13,7 +13,7 @@ use crate::actix::api::read_params::ReadParams;
 use crate::actix::api::CollectionPath;
 use crate::actix::auth::ActixAccess;
 use crate::actix::helpers::{self, get_request_hardware_counter, process_response_error};
-use crate::common::points::do_discover_batch_points;
+use crate::common::query::do_discover_batch_points;
 use crate::settings::ServiceConfig;
 
 #[post("/collections/{name}/points/discover")]
