+++ src/common/mod.rs
@@ -1,22 +1,18 @@
+pub mod auth;
 pub mod collections;
+pub mod debugger;
 pub mod error_reporting;
 pub mod health;
 pub mod helpers;
 pub mod http_client;
+pub mod inference;
 pub mod metrics;
-pub mod points;
+pub mod pyroscope_state;
+pub mod query;
 pub mod snapshots;
 pub mod stacktrace;
+pub mod strings;
 pub mod telemetry;
 pub mod telemetry_ops;
 pub mod telemetry_reporting;
-
-pub mod auth;
-
-pub mod strings;
-
-pub mod debugger;
-
-pub mod inference;
-
-pub mod pyroscope_state;
+pub mod update;
