+++ src/common/query.rs
@@ -1,0 +1,360 @@
+use std::time::Duration;
+
+use api::rest::SearchGroupsRequestInternal;
+use collection::collection::distance_matrix::*;
+use collection::common::batching::batch_requests;
+use collection::grouping::group_by::GroupRequest;
+use collection::operations::consistency_params::ReadConsistency;
+use collection::operations::shard_selector_internal::ShardSelectorInternal;
+use collection::operations::types::*;
+use collection::operations::universal_query::collection_query::*;
+use common::counter::hardware_accumulator::HwMeasurementAcc;
+use segment::types::ScoredPoint;
+use storage::content_manager::errors::StorageError;
+use storage::content_manager::toc::TableOfContent;
+use storage::rbac::Access;
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_core_search_points(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: CoreSearchRequest,
+    read_consistency: Option<ReadConsistency>,
+    shard_selection: ShardSelectorInternal,
+    access: Access,
+    timeout: Option<Duration>,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<Vec<ScoredPoint>, StorageError> {
+    let batch_res = do_core_search_batch_points(
+        toc,
+        collection_name,
+        CoreSearchRequestBatch {
+            searches: vec![request],
+        },
+        read_consistency,
+        shard_selection,
+        access,
+        timeout,
+        hw_measurement_acc,
+    )
+    .await?;
+    batch_res
+        .into_iter()
+        .next()
+        .ok_or_else(|| StorageError::service_error("Empty search result"))
+}
+
+pub async fn do_search_batch_points(
+    toc: &TableOfContent,
+    collection_name: &str,
+    requests: Vec<(CoreSearchRequest, ShardSelectorInternal)>,
+    read_consistency: Option<ReadConsistency>,
+    access: Access,
+    timeout: Option<Duration>,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<Vec<Vec<ScoredPoint>>, StorageError> {
+    let requests = batch_requests::<
+        (CoreSearchRequest, ShardSelectorInternal),
+        ShardSelectorInternal,
+        Vec<CoreSearchRequest>,
+        Vec<_>,
+    >(
+        requests,
+        |(_, shard_selector)| shard_selector,
+        |(request, _), core_reqs| {
+            core_reqs.push(request);
+            Ok(())
+        },
+        |shard_selector, core_requests, res| {
+            if core_requests.is_empty() {
+                return Ok(());
+            }
+
+            let core_batch = CoreSearchRequestBatch {
+                searches: core_requests,
+            };
+
+            let req = toc.core_search_batch(
+                collection_name,
+                core_batch,
+                read_consistency,
+                shard_selector,
+                access.clone(),
+                timeout,
+                hw_measurement_acc.clone(),
+            );
+            res.push(req);
+            Ok(())
+        },
+    )?;
+
+    let results = futures::future::try_join_all(requests).await?;
+    let flatten_results: Vec<Vec<_>> = results.into_iter().flatten().collect();
+    Ok(flatten_results)
+}
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_core_search_batch_points(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: CoreSearchRequestBatch,
+    read_consistency: Option<ReadConsistency>,
+    shard_selection: ShardSelectorInternal,
+    access: Access,
+    timeout: Option<Duration>,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<Vec<Vec<ScoredPoint>>, StorageError> {
+    toc.core_search_batch(
+        collection_name,
+        request,
+        read_consistency,
+        shard_selection,
+        access,
+        timeout,
+        hw_measurement_acc,
+    )
+    .await
+}
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_search_point_groups(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: SearchGroupsRequestInternal,
+    read_consistency: Option<ReadConsistency>,
+    shard_selection: ShardSelectorInternal,
+    access: Access,
+    timeout: Option<Duration>,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<GroupsResult, StorageError> {
+    toc.group(
+        collection_name,
+        GroupRequest::from(request),
+        read_consistency,
+        shard_selection,
+        access,
+        timeout,
+        hw_measurement_acc,
+    )
+    .await
+}
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_recommend_point_groups(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: RecommendGroupsRequestInternal,
+    read_consistency: Option<ReadConsistency>,
+    shard_selection: ShardSelectorInternal,
+    access: Access,
+    timeout: Option<Duration>,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<GroupsResult, StorageError> {
+    toc.group(
+        collection_name,
+        GroupRequest::from(request),
+        read_consistency,
+        shard_selection,
+        access,
+        timeout,
+        hw_measurement_acc,
+    )
+    .await
+}
+
+pub async fn do_discover_batch_points(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: DiscoverRequestBatch,
+    read_consistency: Option<ReadConsistency>,
+    access: Access,
+    timeout: Option<Duration>,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<Vec<Vec<ScoredPoint>>, StorageError> {
+    let requests = request
+        .searches
+        .into_iter()
+        .map(|req| {
+            let shard_selector = match req.shard_key {
+                None => ShardSelectorInternal::All,
+                Some(shard_key) => ShardSelectorInternal::from(shard_key),
+            };
+
+            (req.discover_request, shard_selector)
+        })
+        .collect();
+
+    toc.discover_batch(
+        collection_name,
+        requests,
+        read_consistency,
+        access,
+        timeout,
+        hw_measurement_acc,
+    )
+    .await
+}
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_count_points(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: CountRequestInternal,
+    read_consistency: Option<ReadConsistency>,
+    timeout: Option<Duration>,
+    shard_selection: ShardSelectorInternal,
+    access: Access,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<CountResult, StorageError> {
+    toc.count(
+        collection_name,
+        request,
+        read_consistency,
+        timeout,
+        shard_selection,
+        access,
+        hw_measurement_acc,
+    )
+    .await
+}
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_get_points(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: PointRequestInternal,
+    read_consistency: Option<ReadConsistency>,
+    timeout: Option<Duration>,
+    shard_selection: ShardSelectorInternal,
+    access: Access,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<Vec<RecordInternal>, StorageError> {
+    toc.retrieve(
+        collection_name,
+        request,
+        read_consistency,
+        timeout,
+        shard_selection,
+        access,
+        hw_measurement_acc,
+    )
+    .await
+}
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_scroll_points(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: ScrollRequestInternal,
+    read_consistency: Option<ReadConsistency>,
+    timeout: Option<Duration>,
+    shard_selection: ShardSelectorInternal,
+    access: Access,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<ScrollResult, StorageError> {
+    toc.scroll(
+        collection_name,
+        request,
+        read_consistency,
+        timeout,
+        shard_selection,
+        access,
+        hw_measurement_acc,
+    )
+    .await
+}
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_query_points(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: CollectionQueryRequest,
+    read_consistency: Option<ReadConsistency>,
+    shard_selection: ShardSelectorInternal,
+    access: Access,
+    timeout: Option<Duration>,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<Vec<ScoredPoint>, StorageError> {
+    let requests = vec![(request, shard_selection)];
+    let batch_res = toc
+        .query_batch(
+            collection_name,
+            requests,
+            read_consistency,
+            access,
+            timeout,
+            hw_measurement_acc,
+        )
+        .await?;
+    batch_res
+        .into_iter()
+        .next()
+        .ok_or_else(|| StorageError::service_error("Empty query result"))
+}
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_query_batch_points(
+    toc: &TableOfContent,
+    collection_name: &str,
+    requests: Vec<(CollectionQueryRequest, ShardSelectorInternal)>,
+    read_consistency: Option<ReadConsistency>,
+    access: Access,
+    timeout: Option<Duration>,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<Vec<Vec<ScoredPoint>>, StorageError> {
+    toc.query_batch(
+        collection_name,
+        requests,
+        read_consistency,
+        access,
+        timeout,
+        hw_measurement_acc,
+    )
+    .await
+}
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_query_point_groups(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: CollectionQueryGroupsRequest,
+    read_consistency: Option<ReadConsistency>,
+    shard_selection: ShardSelectorInternal,
+    access: Access,
+    timeout: Option<Duration>,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<GroupsResult, StorageError> {
+    toc.group(
+        collection_name,
+        GroupRequest::from(request),
+        read_consistency,
+        shard_selection,
+        access,
+        timeout,
+        hw_measurement_acc,
+    )
+    .await
+}
+
+#[allow(clippy::too_many_arguments)]
+pub async fn do_search_points_matrix(
+    toc: &TableOfContent,
+    collection_name: &str,
+    request: CollectionSearchMatrixRequest,
+    read_consistency: Option<ReadConsistency>,
+    shard_selection: ShardSelectorInternal,
+    access: Access,
+    timeout: Option<Duration>,
+    hw_measurement_acc: HwMeasurementAcc,
+) -> Result<CollectionSearchMatrixResponse, StorageError> {
+    toc.search_points_matrix(
+        collection_name,
+        request,
+        read_consistency,
+        shard_selection,
+        access,
+        timeout,
+        hw_measurement_acc,
+    )
+    .await
+}
