+++ src/common/update.rs
@@ -1,47 +1,20 @@
 use std::sync::Arc;
-use std::time::Duration;
 
-use api::rest::schema::{PointInsertOperations, PointsBatch, PointsList};
-use api::rest::{SearchGroupsRequestInternal, ShardKeySelector, UpdateVectors};
-use collection::collection::distance_matrix::{
-    CollectionSearchMatrixRequest, CollectionSearchMatrixResponse,
-};
+use api::rest::*;
 use collection::collection::Collection;
-use collection::common::batching::batch_requests;
-use collection::grouping::group_by::GroupRequest;
-use collection::operations::consistency_params::ReadConsistency;
-use collection::operations::payload_ops::{
-    DeletePayload, DeletePayloadOp, PayloadOps, SetPayload, SetPayloadOp,
-};
-use collection::operations::point_ops::{
-    FilterSelector, PointIdsList, PointInsertOperationsInternal, PointOperations, PointsSelector,
-    WriteOrdering,
-};
+use collection::operations::payload_ops::*;
+use collection::operations::point_ops::*;
 use collection::operations::shard_selector_internal::ShardSelectorInternal;
-use collection::operations::types::{
-    CollectionError, CoreSearchRequest, CoreSearchRequestBatch, CountRequestInternal, CountResult,
-    DiscoverRequestBatch, GroupsResult, PointRequestInternal, RecommendGroupsRequestInternal,
-    RecordInternal, ScrollRequestInternal, ScrollResult, UpdateResult,
-};
-use collection::operations::universal_query::collection_query::{
-    CollectionQueryGroupsRequest, CollectionQueryRequest,
-};
-use collection::operations::vector_ops::{DeleteVectors, UpdateVectorsOp, VectorOperations};
-use collection::operations::verification::{
-    new_unchecked_verification_pass, StrictModeVerification,
-};
-use collection::operations::{
-    ClockTag, CollectionUpdateOperations, CreateIndex, FieldIndexOperations, OperationWithClockTag,
-};
+use collection::operations::types::{CollectionError, UpdateResult};
+use collection::operations::vector_ops::*;
+use collection::operations::verification::*;
+use collection::operations::*;
 use collection::shards::shard::ShardId;
-use common::counter::hardware_accumulator::HwMeasurementAcc;
 use schemars::JsonSchema;
 use segment::json_path::JsonPath;
-use segment::types::{PayloadFieldSchema, PayloadKeyType, ScoredPoint, StrictModeConfig};
+use segment::types::{PayloadFieldSchema, PayloadKeyType, StrictModeConfig};
 use serde::{Deserialize, Serialize};
-use storage::content_manager::collection_meta_ops::{
-    CollectionMetaOperations, CreatePayloadIndex, DropPayloadIndex,
-};
+use storage::content_manager::collection_meta_ops::*;
 use storage::content_manager::errors::StorageError;
 use storage::content_manager::toc::TableOfContent;
 use storage::dispatcher::Dispatcher;
@@ -49,16 +22,112 @@ use storage::rbac::Access;
 use validator::Validate;
 
 use crate::common::inference::service::InferenceType;
-use crate::common::inference::update_requests::{
-    convert_batch, convert_point_struct, convert_point_vectors,
-};
+use crate::common::inference::update_requests::*;
 use crate::common::inference::InferenceToken;
 
-#[derive(Debug, Deserialize, Serialize, JsonSchema, Validate)]
-pub struct CreateFieldIndex {
-    pub field_name: PayloadKeyType,
-    #[serde(alias = "field_type")]
-    pub field_schema: Option<PayloadFieldSchema>,
+#[derive(Deserialize, Serialize, JsonSchema, Validate)]
+pub struct UpdateOperations {
+    pub operations: Vec<UpdateOperation>,
+}
+
+#[derive(Deserialize, Serialize, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+#[serde(untagged)]
+pub enum UpdateOperation {
+    Upsert(UpsertOperation),
+    Delete(DeleteOperation),
+    SetPayload(SetPayloadOperation),
+    OverwritePayload(OverwritePayloadOperation),
+    DeletePayload(DeletePayloadOperation),
+    ClearPayload(ClearPayloadOperation),
+    UpdateVectors(UpdateVectorsOperation),
+    DeleteVectors(DeleteVectorsOperation),
+}
+
+impl Validate for UpdateOperation {
+    fn validate(&self) -> Result<(), validator::ValidationErrors> {
+        match self {
+            UpdateOperation::Upsert(op) => op.validate(),
+            UpdateOperation::Delete(op) => op.validate(),
+            UpdateOperation::SetPayload(op) => op.validate(),
+            UpdateOperation::OverwritePayload(op) => op.validate(),
+            UpdateOperation::DeletePayload(op) => op.validate(),
+            UpdateOperation::ClearPayload(op) => op.validate(),
+            UpdateOperation::UpdateVectors(op) => op.validate(),
+            UpdateOperation::DeleteVectors(op) => op.validate(),
+        }
+    }
+}
+
+impl StrictModeVerification for UpdateOperation {
+    fn query_limit(&self) -> Option<usize> {
+        None
+    }
+
+    fn indexed_filter_read(&self) -> Option<&segment::types::Filter> {
+        None
+    }
+
+    fn indexed_filter_write(&self) -> Option<&segment::types::Filter> {
+        None
+    }
+
+    fn request_exact(&self) -> Option<bool> {
+        None
+    }
+
+    fn request_search_params(&self) -> Option<&segment::types::SearchParams> {
+        None
+    }
+
+    async fn check_strict_mode(
+        &self,
+        collection: &Collection,
+        strict_mode_config: &StrictModeConfig,
+    ) -> Result<(), CollectionError> {
+        match self {
+            UpdateOperation::Upsert(op) => {
+                op.upsert
+                    .check_strict_mode(collection, strict_mode_config)
+                    .await
+            }
+            UpdateOperation::Delete(op) => {
+                op.delete
+                    .check_strict_mode(collection, strict_mode_config)
+                    .await
+            }
+            UpdateOperation::SetPayload(op) => {
+                op.set_payload
+                    .check_strict_mode(collection, strict_mode_config)
+                    .await
+            }
+            UpdateOperation::OverwritePayload(op) => {
+                op.overwrite_payload
+                    .check_strict_mode(collection, strict_mode_config)
+                    .await
+            }
+            UpdateOperation::DeletePayload(op) => {
+                op.delete_payload
+                    .check_strict_mode(collection, strict_mode_config)
+                    .await
+            }
+            UpdateOperation::ClearPayload(op) => {
+                op.clear_payload
+                    .check_strict_mode(collection, strict_mode_config)
+                    .await
+            }
+            UpdateOperation::UpdateVectors(op) => {
+                op.update_vectors
+                    .check_strict_mode(collection, strict_mode_config)
+                    .await
+            }
+            UpdateOperation::DeleteVectors(op) => {
+                op.delete_vectors
+                    .check_strict_mode(collection, strict_mode_config)
+                    .await
+            }
+        }
+    }
 }
 
 #[derive(Deserialize, Serialize, JsonSchema, Validate)]
@@ -109,148 +178,11 @@ pub struct DeleteVectorsOperation {
     delete_vectors: DeleteVectors,
 }
 
-#[derive(Deserialize, Serialize, JsonSchema)]
-#[serde(rename_all = "snake_case")]
-#[serde(untagged)]
-pub enum UpdateOperation {
-    Upsert(UpsertOperation),
-    Delete(DeleteOperation),
-    SetPayload(SetPayloadOperation),
-    OverwritePayload(OverwritePayloadOperation),
-    DeletePayload(DeletePayloadOperation),
-    ClearPayload(ClearPayloadOperation),
-    UpdateVectors(UpdateVectorsOperation),
-    DeleteVectors(DeleteVectorsOperation),
-}
-
-#[derive(Deserialize, Serialize, JsonSchema, Validate)]
-pub struct UpdateOperations {
-    pub operations: Vec<UpdateOperation>,
-}
-
-impl Validate for UpdateOperation {
-    fn validate(&self) -> Result<(), validator::ValidationErrors> {
-        match self {
-            UpdateOperation::Upsert(op) => op.validate(),
-            UpdateOperation::Delete(op) => op.validate(),
-            UpdateOperation::SetPayload(op) => op.validate(),
-            UpdateOperation::OverwritePayload(op) => op.validate(),
-            UpdateOperation::DeletePayload(op) => op.validate(),
-            UpdateOperation::ClearPayload(op) => op.validate(),
-            UpdateOperation::UpdateVectors(op) => op.validate(),
-            UpdateOperation::DeleteVectors(op) => op.validate(),
-        }
-    }
-}
-
-impl StrictModeVerification for UpdateOperation {
-    fn query_limit(&self) -> Option<usize> {
-        None
-    }
-
-    fn indexed_filter_read(&self) -> Option<&segment::types::Filter> {
-        None
-    }
-
-    fn indexed_filter_write(&self) -> Option<&segment::types::Filter> {
-        None
-    }
-
-    fn request_exact(&self) -> Option<bool> {
-        None
-    }
-
-    fn request_search_params(&self) -> Option<&segment::types::SearchParams> {
-        None
-    }
-
-    async fn check_strict_mode(
-        &self,
-        collection: &Collection,
-        strict_mode_config: &StrictModeConfig,
-    ) -> Result<(), CollectionError> {
-        match self {
-            UpdateOperation::Delete(delete_op) => {
-                delete_op
-                    .delete
-                    .check_strict_mode(collection, strict_mode_config)
-                    .await
-            }
-            UpdateOperation::SetPayload(set_payload) => {
-                set_payload
-                    .set_payload
-                    .check_strict_mode(collection, strict_mode_config)
-                    .await
-            }
-            UpdateOperation::OverwritePayload(overwrite_payload) => {
-                overwrite_payload
-                    .overwrite_payload
-                    .check_strict_mode(collection, strict_mode_config)
-                    .await
-            }
-            UpdateOperation::DeletePayload(delete_payload) => {
-                delete_payload
-                    .delete_payload
-                    .check_strict_mode(collection, strict_mode_config)
-                    .await
-            }
-            UpdateOperation::ClearPayload(clear_payload) => {
-                clear_payload
-                    .clear_payload
-                    .check_strict_mode(collection, strict_mode_config)
-                    .await
-            }
-            UpdateOperation::DeleteVectors(delete_op) => {
-                delete_op
-                    .delete_vectors
-                    .check_strict_mode(collection, strict_mode_config)
-                    .await
-            }
-            UpdateOperation::Upsert(upsert_op) => {
-                upsert_op
-                    .upsert
-                    .check_strict_mode(collection, strict_mode_config)
-                    .await
-            }
-            UpdateOperation::UpdateVectors(update) => {
-                update
-                    .update_vectors
-                    .check_strict_mode(collection, strict_mode_config)
-                    .await
-            }
-        }
-    }
-}
-
-/// Converts a pair of parameters into a shard selector
-/// suitable for update operations.
-///
-/// The key difference from selector for search operations is that
-/// empty shard selector in case of update means default shard,
-/// while empty shard selector in case of search means all shards.
-///
-/// Parameters:
-/// - shard_selection: selection of the exact shard ID, always have priority over shard_key
-/// - shard_key: selection of the shard key, can be a single key or a list of keys
-///
-/// Returns:
-/// - ShardSelectorInternal - resolved shard selector
-fn get_shard_selector_for_update(
-    shard_selection: Option<ShardId>,
-    shard_key: Option<ShardKeySelector>,
-) -> ShardSelectorInternal {
-    match (shard_selection, shard_key) {
-        (Some(shard_selection), None) => ShardSelectorInternal::ShardId(shard_selection),
-        (Some(shard_selection), Some(_)) => {
-            debug_assert!(
-                false,
-                "Shard selection and shard key are mutually exclusive"
-            );
-            ShardSelectorInternal::ShardId(shard_selection)
-        }
-        (None, Some(shard_key)) => ShardSelectorInternal::from(shard_key),
-        (None, None) => ShardSelectorInternal::Empty,
-    }
+#[derive(Debug, Deserialize, Serialize, JsonSchema, Validate)]
+pub struct CreateFieldIndex {
+    pub field_name: PayloadKeyType,
+    #[serde(alias = "field_type")]
+    pub field_schema: Option<PayloadFieldSchema>,
 }
 
 #[allow(clippy::too_many_arguments)]
@@ -712,41 +644,6 @@ pub async fn do_batch_update_points(
 }
 
 #[allow(clippy::too_many_arguments)]
-pub async fn do_create_index_internal(
-    toc: Arc<TableOfContent>,
-    collection_name: String,
-    field_name: PayloadKeyType,
-    field_schema: Option<PayloadFieldSchema>,
-    clock_tag: Option<ClockTag>,
-    shard_selection: Option<ShardId>,
-    wait: bool,
-    ordering: WriteOrdering,
-) -> Result<UpdateResult, StorageError> {
-    let collection_operation = CollectionUpdateOperations::FieldIndexOperation(
-        FieldIndexOperations::CreateIndex(CreateIndex {
-            field_name,
-            field_schema,
-        }),
-    );
-
-    let shard_selector = if let Some(shard_selection) = shard_selection {
-        ShardSelectorInternal::ShardId(shard_selection)
-    } else {
-        ShardSelectorInternal::All
-    };
-
-    toc.update(
-        &collection_name,
-        OperationWithClockTag::new(collection_operation, clock_tag),
-        wait,
-        ordering,
-        shard_selector,
-        Access::full("Internal API"),
-    )
-    .await
-}
-
-#[allow(clippy::too_many_arguments)]
 pub async fn do_create_index(
     dispatcher: Arc<Dispatcher>,
     collection_name: String,
@@ -802,17 +699,21 @@ pub async fn do_create_index(
 }
 
 #[allow(clippy::too_many_arguments)]
-pub async fn do_delete_index_internal(
+pub async fn do_create_index_internal(
     toc: Arc<TableOfContent>,
     collection_name: String,
-    index_name: JsonPath,
+    field_name: PayloadKeyType,
+    field_schema: Option<PayloadFieldSchema>,
     clock_tag: Option<ClockTag>,
     shard_selection: Option<ShardId>,
     wait: bool,
     ordering: WriteOrdering,
 ) -> Result<UpdateResult, StorageError> {
     let collection_operation = CollectionUpdateOperations::FieldIndexOperation(
-        FieldIndexOperations::DeleteIndex(index_name),
+        FieldIndexOperations::CreateIndex(CreateIndex {
+            field_name,
+            field_schema,
+        }),
     );
 
     let shard_selector = if let Some(shard_selection) = shard_selection {
@@ -876,346 +777,63 @@ pub async fn do_delete_index(
 }
 
 #[allow(clippy::too_many_arguments)]
-pub async fn do_core_search_points(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: CoreSearchRequest,
-    read_consistency: Option<ReadConsistency>,
-    shard_selection: ShardSelectorInternal,
-    access: Access,
-    timeout: Option<Duration>,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<Vec<ScoredPoint>, StorageError> {
-    let batch_res = do_core_search_batch_points(
-        toc,
-        collection_name,
-        CoreSearchRequestBatch {
-            searches: vec![request],
-        },
-        read_consistency,
-        shard_selection,
-        access,
-        timeout,
-        hw_measurement_acc,
+pub async fn do_delete_index_internal(
+    toc: Arc<TableOfContent>,
+    collection_name: String,
+    index_name: JsonPath,
+    clock_tag: Option<ClockTag>,
+    shard_selection: Option<ShardId>,
+    wait: bool,
+    ordering: WriteOrdering,
+) -> Result<UpdateResult, StorageError> {
+    let collection_operation = CollectionUpdateOperations::FieldIndexOperation(
+        FieldIndexOperations::DeleteIndex(index_name),
+    );
+
+    let shard_selector = if let Some(shard_selection) = shard_selection {
+        ShardSelectorInternal::ShardId(shard_selection)
+    } else {
+        ShardSelectorInternal::All
+    };
+
+    toc.update(
+        &collection_name,
+        OperationWithClockTag::new(collection_operation, clock_tag),
+        wait,
+        ordering,
+        shard_selector,
+        Access::full("Internal API"),
     )
-    .await?;
-    batch_res
-        .into_iter()
-        .next()
-        .ok_or_else(|| StorageError::service_error("Empty search result"))
+    .await
 }
 
-pub async fn do_search_batch_points(
-    toc: &TableOfContent,
-    collection_name: &str,
-    requests: Vec<(CoreSearchRequest, ShardSelectorInternal)>,
-    read_consistency: Option<ReadConsistency>,
-    access: Access,
-    timeout: Option<Duration>,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<Vec<Vec<ScoredPoint>>, StorageError> {
-    let requests = batch_requests::<
-        (CoreSearchRequest, ShardSelectorInternal),
-        ShardSelectorInternal,
-        Vec<CoreSearchRequest>,
-        Vec<_>,
-    >(
-        requests,
-        |(_, shard_selector)| shard_selector,
-        |(request, _), core_reqs| {
-            core_reqs.push(request);
-            Ok(())
-        },
-        |shard_selector, core_requests, res| {
-            if core_requests.is_empty() {
-                return Ok(());
-            }
-
-            let core_batch = CoreSearchRequestBatch {
-                searches: core_requests,
-            };
-
-            let req = toc.core_search_batch(
-                collection_name,
-                core_batch,
-                read_consistency,
-                shard_selector,
-                access.clone(),
-                timeout,
-                hw_measurement_acc.clone(),
+/// Converts a pair of parameters into a shard selector
+/// suitable for update operations.
+///
+/// The key difference from selector for search operations is that
+/// empty shard selector in case of update means default shard,
+/// while empty shard selector in case of search means all shards.
+///
+/// Parameters:
+/// - shard_selection: selection of the exact shard ID, always have priority over shard_key
+/// - shard_key: selection of the shard key, can be a single key or a list of keys
+///
+/// Returns:
+/// - ShardSelectorInternal - resolved shard selector
+fn get_shard_selector_for_update(
+    shard_selection: Option<ShardId>,
+    shard_key: Option<ShardKeySelector>,
+) -> ShardSelectorInternal {
+    match (shard_selection, shard_key) {
+        (Some(shard_selection), None) => ShardSelectorInternal::ShardId(shard_selection),
+        (Some(shard_selection), Some(_)) => {
+            debug_assert!(
+                false,
+                "Shard selection and shard key are mutually exclusive"
             );
-            res.push(req);
-            Ok(())
-        },
-    )?;
-
-    let results = futures::future::try_join_all(requests).await?;
-    let flatten_results: Vec<Vec<_>> = results.into_iter().flatten().collect();
-    Ok(flatten_results)
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn do_core_search_batch_points(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: CoreSearchRequestBatch,
-    read_consistency: Option<ReadConsistency>,
-    shard_selection: ShardSelectorInternal,
-    access: Access,
-    timeout: Option<Duration>,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<Vec<Vec<ScoredPoint>>, StorageError> {
-    toc.core_search_batch(
-        collection_name,
-        request,
-        read_consistency,
-        shard_selection,
-        access,
-        timeout,
-        hw_measurement_acc,
-    )
-    .await
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn do_search_point_groups(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: SearchGroupsRequestInternal,
-    read_consistency: Option<ReadConsistency>,
-    shard_selection: ShardSelectorInternal,
-    access: Access,
-    timeout: Option<Duration>,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<GroupsResult, StorageError> {
-    toc.group(
-        collection_name,
-        GroupRequest::from(request),
-        read_consistency,
-        shard_selection,
-        access,
-        timeout,
-        hw_measurement_acc,
-    )
-    .await
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn do_recommend_point_groups(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: RecommendGroupsRequestInternal,
-    read_consistency: Option<ReadConsistency>,
-    shard_selection: ShardSelectorInternal,
-    access: Access,
-    timeout: Option<Duration>,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<GroupsResult, StorageError> {
-    toc.group(
-        collection_name,
-        GroupRequest::from(request),
-        read_consistency,
-        shard_selection,
-        access,
-        timeout,
-        hw_measurement_acc,
-    )
-    .await
-}
-
-pub async fn do_discover_batch_points(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: DiscoverRequestBatch,
-    read_consistency: Option<ReadConsistency>,
-    access: Access,
-    timeout: Option<Duration>,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<Vec<Vec<ScoredPoint>>, StorageError> {
-    let requests = request
-        .searches
-        .into_iter()
-        .map(|req| {
-            let shard_selector = match req.shard_key {
-                None => ShardSelectorInternal::All,
-                Some(shard_key) => ShardSelectorInternal::from(shard_key),
-            };
-
-            (req.discover_request, shard_selector)
-        })
-        .collect();
-
-    toc.discover_batch(
-        collection_name,
-        requests,
-        read_consistency,
-        access,
-        timeout,
-        hw_measurement_acc,
-    )
-    .await
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn do_count_points(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: CountRequestInternal,
-    read_consistency: Option<ReadConsistency>,
-    timeout: Option<Duration>,
-    shard_selection: ShardSelectorInternal,
-    access: Access,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<CountResult, StorageError> {
-    toc.count(
-        collection_name,
-        request,
-        read_consistency,
-        timeout,
-        shard_selection,
-        access,
-        hw_measurement_acc,
-    )
-    .await
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn do_get_points(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: PointRequestInternal,
-    read_consistency: Option<ReadConsistency>,
-    timeout: Option<Duration>,
-    shard_selection: ShardSelectorInternal,
-    access: Access,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<Vec<RecordInternal>, StorageError> {
-    toc.retrieve(
-        collection_name,
-        request,
-        read_consistency,
-        timeout,
-        shard_selection,
-        access,
-        hw_measurement_acc,
-    )
-    .await
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn do_scroll_points(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: ScrollRequestInternal,
-    read_consistency: Option<ReadConsistency>,
-    timeout: Option<Duration>,
-    shard_selection: ShardSelectorInternal,
-    access: Access,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<ScrollResult, StorageError> {
-    toc.scroll(
-        collection_name,
-        request,
-        read_consistency,
-        timeout,
-        shard_selection,
-        access,
-        hw_measurement_acc,
-    )
-    .await
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn do_query_points(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: CollectionQueryRequest,
-    read_consistency: Option<ReadConsistency>,
-    shard_selection: ShardSelectorInternal,
-    access: Access,
-    timeout: Option<Duration>,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<Vec<ScoredPoint>, StorageError> {
-    let requests = vec![(request, shard_selection)];
-    let batch_res = toc
-        .query_batch(
-            collection_name,
-            requests,
-            read_consistency,
-            access,
-            timeout,
-            hw_measurement_acc,
-        )
-        .await?;
-    batch_res
-        .into_iter()
-        .next()
-        .ok_or_else(|| StorageError::service_error("Empty query result"))
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn do_query_batch_points(
-    toc: &TableOfContent,
-    collection_name: &str,
-    requests: Vec<(CollectionQueryRequest, ShardSelectorInternal)>,
-    read_consistency: Option<ReadConsistency>,
-    access: Access,
-    timeout: Option<Duration>,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<Vec<Vec<ScoredPoint>>, StorageError> {
-    toc.query_batch(
-        collection_name,
-        requests,
-        read_consistency,
-        access,
-        timeout,
-        hw_measurement_acc,
-    )
-    .await
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn do_query_point_groups(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: CollectionQueryGroupsRequest,
-    read_consistency: Option<ReadConsistency>,
-    shard_selection: ShardSelectorInternal,
-    access: Access,
-    timeout: Option<Duration>,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<GroupsResult, StorageError> {
-    toc.group(
-        collection_name,
-        GroupRequest::from(request),
-        read_consistency,
-        shard_selection,
-        access,
-        timeout,
-        hw_measurement_acc,
-    )
-    .await
-}
-
-#[allow(clippy::too_many_arguments)]
-pub async fn do_search_points_matrix(
-    toc: &TableOfContent,
-    collection_name: &str,
-    request: CollectionSearchMatrixRequest,
-    read_consistency: Option<ReadConsistency>,
-    shard_selection: ShardSelectorInternal,
-    access: Access,
-    timeout: Option<Duration>,
-    hw_measurement_acc: HwMeasurementAcc,
-) -> Result<CollectionSearchMatrixResponse, StorageError> {
-    toc.search_points_matrix(
-        collection_name,
-        request,
-        read_consistency,
-        shard_selection,
-        access,
-        timeout,
-        hw_measurement_acc,
-    )
-    .await
+            ShardSelectorInternal::ShardId(shard_selection)
+        }
+        (None, Some(shard_key)) => ShardSelectorInternal::from(shard_key),
+        (None, None) => ShardSelectorInternal::Empty,
+    }
 }
